✨ NEW: renamed useLogin to useAuth
